flytt filter-logikk til egen pakke
