+++ frontend/mr-admin-flate/src/api/atoms.ts
@@ -1,4 +1,4 @@
-import { ARRANGORER_PAGE_SIZE, AVTALE_PAGE_SIZE, PAGE_SIZE } from "@/constants";
+import { ARRANGORER_PAGE_SIZE, AVTALE_PAGE_SIZE } from "@/constants";
 import { SortState } from "@navikt/ds-react";
 import { atom, WritableAtom } from "jotai";
 import {
@@ -10,17 +10,14 @@ import {
 import {
   Avtalestatus,
   Avtaletype,
-  GjennomforingStatus,
-  NavEnhet,
   OppgaveType,
   SorteringArrangorer,
   SorteringAvtaler,
-  SorteringGjennomforinger,
   SorteringTiltakstyper,
 } from "@mr/api-client-v2";
 import { z, ZodType } from "zod";
 
-function createSorteringProps(sortItems: z.ZodType) {
+export function createSorteringProps(sortItems: z.ZodType) {
   return z.object({
     tableSort: z.custom<SortState>(),
     sortString: sortItems,
@@ -59,184 +56,6 @@ export const tiltakstypeFilterAtom = atomWithStorage<TiltakstypeFilterType>(
   tiltakstypeFilterStorage,
   { getOnInit: true },
 );
-
-export const GjennomforingFilterSchema = z.object({
-  search: z.string(),
-  navEnheter: z.custom<NavEnhet>().array(),
-  tiltakstyper: z.string().array(),
-  statuser: z.custom<GjennomforingStatus>().array(),
-  sortering: createSorteringProps(z.custom<SorteringGjennomforinger>()),
-  avtale: z.string(),
-  arrangorer: z.string().array(),
-  visMineGjennomforinger: z.boolean(),
-  publisert: z.string().array(),
-  page: z.number(),
-  pageSize: z.number(),
-});
-
-export type GjennomforingFilterType = z.infer<typeof GjennomforingFilterSchema>;
-
-export const defaultGjennomforingFilter: GjennomforingFilterType = {
-  search: "",
-  navEnheter: [],
-  tiltakstyper: [],
-  statuser: [GjennomforingStatus.GJENNOMFORES],
-  sortering: {
-    sortString: SorteringGjennomforinger.NAVN_ASCENDING,
-    tableSort: {
-      orderBy: "navn",
-      direction: "ascending",
-    },
-  },
-  avtale: "",
-  arrangorer: [],
-  publisert: [],
-  visMineGjennomforinger: false,
-  page: 1,
-  pageSize: PAGE_SIZE,
-  // lagretFilterIdValgt: undefined,
-};
-
-const gjennomforingFilterStorage = withStorageValidator(
-  createFilterValidator(GjennomforingFilterSchema),
-)(createJSONStorage(() => sessionStorage));
-
-export const gjennomforingfilterAtom = atomWithStorage<GjennomforingFilterType>(
-  "gjennomforing-filter",
-  defaultGjennomforingFilter,
-  gjennomforingFilterStorage,
-  { getOnInit: true },
-);
-
-const gjennomforingerForAvtaleFilterAtomFamily = atomFamily<
-  GjennomforingFilterType,
-  WritableAtom<GjennomforingFilterType, [newValue: GjennomforingFilterType], void>
->(
-  (filter: GjennomforingFilterType) => {
-    return atomWithStorage(
-      `gjennomforing-filter-${filter.avtale}`,
-      filter,
-      createJSONStorage(() => sessionStorage),
-    );
-  },
-  (a, b) => a.avtale === b.avtale,
-);
-
-export function getGjennomforingerForAvtaleFilterAtom(avtaleId: string) {
-  const defaultFilterValue = { ...defaultGjennomforingFilter, avtale: avtaleId };
-  const filterAtom = gjennomforingerForAvtaleFilterAtomFamily(defaultFilterValue);
-  return { defaultFilterValue, filterAtom };
-}
-
-export type FilterState<T> = {
-  filter: FilterEntry<T>;
-  defaultFilter: FilterEntry<T>;
-  isInitialized: boolean;
-};
-
-export type FilterEntry<T> = {
-  id?: string;
-  values: T;
-};
-
-export type FilterAction<T> =
-  | { type: "UPDATE_DEFAULT"; payload?: FilterEntry<T> }
-  | { type: "RESET_TO_DEFAULT" }
-  | { type: "SELECT_FILTER"; payload: FilterEntry<T> }
-  | { type: "SET_FILTER"; payload: T }
-  | { type: "UPDATE_FILTER"; payload: Partial<T> };
-
-export function createFilterStateAtom<T extends object>(
-  storageKey: string,
-  fallbackFilter: T,
-  validator: (value: unknown) => value is T,
-) {
-  const initialState = {
-    isInitialized: false,
-    defaultFilter: { id: undefined, values: fallbackFilter },
-    filter: { values: fallbackFilter },
-  };
-
-  // Create storage
-  const filterStateStorage = withStorageValidator<FilterState<T>>(
-    (value: unknown): value is FilterState<T> => {
-      if (!value || typeof value !== "object") return false;
-      const state = value as any;
-      return (
-        state.filter &&
-        validator(state.filter.values) &&
-        (typeof state.filter.id === "string" || state.filter.id === undefined) &&
-        (state.defaultFilter === undefined || validator(state.defaultFilter.values)) &&
-        typeof state.isInitialized === "boolean"
-      );
-    },
-  )(createJSONStorage(() => sessionStorage));
-
-  function filterReducer(state: FilterState<T>, action: FilterAction<T>): FilterState<T> {
-    switch (action.type) {
-      case "UPDATE_DEFAULT": {
-        const defaultFilter = action.payload ?? initialState.defaultFilter;
-        if (!state.isInitialized) {
-          return {
-            ...state,
-            isInitialized: true,
-            filter: defaultFilter,
-            defaultFilter,
-          };
-        }
-
-        return { ...state, defaultFilter };
-      }
-
-      case "RESET_TO_DEFAULT":
-        return {
-          ...state,
-          filter: state.defaultFilter,
-        };
-
-      case "SET_FILTER":
-        return {
-          ...state,
-          filter: {
-            id: undefined,
-            values: action.payload,
-          },
-        };
-
-      case "UPDATE_FILTER":
-        return {
-          ...state,
-          filter: {
-            id: undefined,
-            values: { ...state.filter.values, ...action.payload },
-          },
-        };
-
-      case "SELECT_FILTER":
-        return {
-          ...state,
-          filter: action.payload,
-        };
-
-      default:
-        return state;
-    }
-  }
-
-  const filterManagerAtom = atomWithStorage<FilterState<T>>(
-    storageKey,
-    initialState,
-    filterStateStorage,
-    { getOnInit: true },
-  );
-
-  return atom(
-    (get) => get(filterManagerAtom),
-    (get, set, action: FilterAction<T>) => {
-      set(filterManagerAtom, filterReducer(get(filterManagerAtom), action));
-    },
-  );
-}
 
 export const AvtaleFilterSchema = z.object({
   sok: z.string(),
@@ -370,6 +189,5 @@ export const avtaleDetaljerTabAtom = atom<
   "detaljer" | "okonomi" | "personvern" | "redaksjonelt-innhold"
 >("detaljer");
 
-export const gjennomforingFilterAccordionAtom = atom<string[]>(["navEnhet"]);
 export const avtaleFilterAccordionAtom = atom<string[]>(["region"]);
 export const oppgaverFilterAccordionAtom = atom<string[]>(["type", "regioner"]);
