+++ frontend/mr-admin-flate/src/api/gjennomforing/useAdminGjennomforinger.ts
@@ -1,8 +1,8 @@
 import { useApiQuery, useDebounce } from "@mr/frontend-common";
 import { QueryKeys } from "@/api/QueryKeys";
-import { GjennomforingFilterType } from "../atoms";
 import { type GetGjennomforingerData, GjennomforingerService } from "@mr/api-client-v2";
 import { getPublisertStatus } from "@/utils/Utils";
+import { GjennomforingFilterType } from "@/pages/gjennomforing/filter";
 
 export function useAdminGjennomforinger(filter: Partial<GjennomforingFilterType>) {
   const debouncedSok = useDebounce(filter.search?.trim(), 300);
