+++ frontend/mr-admin-flate/src/components/filter/GjennomforingFilterTags.tsx
@@ -1,10 +1,10 @@
 import { ArrangorTil } from "@mr/api-client-v2";
-import { GjennomforingFilterType } from "@/api/atoms";
 import { useTiltakstyper } from "@/api/tiltakstyper/useTiltakstyper";
 import { useArrangorer } from "@/api/arrangor/useArrangorer";
 import { addOrRemove } from "@mr/frontend-common/utils/utils";
 import { TILTAKSGJENNOMFORING_STATUS_OPTIONS } from "@/utils/filterUtils";
 import { FilterTag, FilterTagsContainer, NavEnhetFilterTag } from "@mr/frontend-common";
+import { GjennomforingFilterType } from "@/pages/gjennomforing/filter";
 
 interface Props {
   filter: GjennomforingFilterType;
